Refactors and fixes in user management Functions handlers (#138)
The Functions code for user management would under the right circumstances result in a hung request. This sparked a refactor of the code to consistently apply async-await style (over continuation passing style), which made it easier to check for (and fix) branches where the HTTP response was not being correctly closed.